chore(taro-router-rn): move getOpenerEventChannel
