+++ packages/taro-router-rn/src/rootNavigation.ts
@@ -1,11 +1,11 @@
 // RootNavigation.js
 import * as React from 'react'
-import { DeviceEventEmitter } from 'react-native'
 import { camelCase } from 'lodash'
 import { parseUrl } from 'query-string'
 import { StackActions, NavigationContainerRef } from '@react-navigation/native'
 import { getTabBarPages } from './utils/index'
 import { CallbackResult, BaseOption } from './utils/types'
+import { getOpenerEventChannel } from './getOpenerEventChannel'
 
 type NavigateMethod = 'navigateTo' | 'redirectTo' | 'navigateBack' | 'switchTab' | 'reLaunch'
 
@@ -20,64 +20,6 @@ interface NavigateBackOption extends BaseOption {
 interface NavigateRef extends NavigationContainerRef {
   setOptions: (obj: any) => void
 }
-
-let emitterList: any[] = []
-
-function getOpenerEventChannel () {
-  return {
-    emit (eventName, ...args) {
-      DeviceEventEmitter.emit(eventName, ...args)
-    },
-    on (eventName, callback) {
-      emitterList.push({
-        eventName,
-        cbName: callback.name,
-        once: false,
-        emitter: DeviceEventEmitter.addListener(eventName, res => callback(res))
-      })
-    },
-    once (eventName, callback) {
-      emitterList.push({
-        eventName,
-        cbName: callback.name,
-        once: true,
-        emitter: DeviceEventEmitter.addListener(eventName, res => {
-          callback(res)
-          emitterList.forEach(item => {
-            if (item.eventName === eventName && item.once && item.emitter) {
-              item.emitter.remove()
-              item.emitter = null
-            }
-          })
-        })
-      })
-    },
-    off (eventName, callback) {
-      if (callback && typeof callback === 'function') {
-        for (let i = 0; i < emitterList.length; i++) {
-          const item = emitterList[i]
-          if (item.emitter && eventName === item.eventName && callback.name && callback.name === item.cbName) {
-            item.emitter.remove()
-            item.emitter = null
-            break
-          }
-        }
-      } else {
-        emitterList.forEach((item) => {
-          if (eventName === item.eventName && item.emitter) {
-            item.emitter.remove()
-            item.emitter = null
-          }
-        })
-      }
-      emitterList = emitterList.filter(item => item.emitter)
-    }
-  }
-}
-// 挂载至原型，从而this.getOpenerEventChannel()方式使用,对标微信小程序相关方法
-Object.assign(React.Component.prototype, {
-  getOpenerEventChannel
-})
 
 export const isReadyRef = React.createRef()
